+++ src/components/ResourcesEditor/ResourcesEditor.tsx
@@ -1,6 +1,5 @@
 import { StandardEditorProps } from '@grafana/data';
-import { config } from '@grafana/runtime';
-import { Alert, Button, Icon, InlineField, InlineFieldRow, Input, useStyles2 } from '@grafana/ui';
+import { Button, Icon, InlineField, InlineFieldRow, Input, useStyles2 } from '@grafana/ui';
 import { Collapse } from '@volkovlabs/components';
 import React, { useCallback, useState } from 'react';
 import {
@@ -20,7 +19,7 @@ import { getStyles } from './ResourceEditor.styles';
 /**
  * Properties
  */
-type Props = StandardEditorProps<Resource[], null, PanelOptions>;
+type Props = StandardEditorProps<Resource[], PanelOptions>;
 
 /**
  * Reorder
@@ -128,17 +127,6 @@ export const ResourcesEditor: React.FC<Props> = ({ value, onChange }) => {
     },
     [items, onChangeItems]
   );
-
-  /**
-   * Disabled sanitize
-   */
-  if (!config?.disableSanitizeHtml) {
-    return (
-      <Alert title="Unavailable" severity="info" data-testid={TEST_IDS.resourcesEditor.infoMessage}>
-        Please disable sanitize HTML to use external resources.
-      </Alert>
-    );
-  }
 
   return (
     <>